Update type based on BVH::LevelInside

+++ src/geocel/vg/VecgeomTypes.hh
@@ -57,7 +57,7 @@ namespace celeritas
 //---------------------------------------------------------------------------//
 
 using VgSurfaceInt = long;
-using VgPlacedVolumeInt = int;
+using VgPlacedVolumeInt = long;
 using vg_real_type = VECGEOM_PRECISION_NAMESPACE::Precision;
 
 #if defined(VECGEOM_BVH_SINGLE) || defined(__DOXYGEN__)
